requirements.txt
fastapi==0.114.0
uvicorn==0.23.2
python-telegram-bot==20.6
openai==1.10.2
